+++ constants/__init__.py
@@ -32,7 +32,7 @@ V1_MODEL_ID = "v1"
 COMPETITION_SCHEDULE: List[CompetitionParameters] = [
     CompetitionParameters(
         reward_percentage=1.0,
-        competition_id="v3",
+        competition_id="v4",
     ),
 ]
 ORIGINAL_COMPETITION_ID = O1_MODEL_ID
@@ -83,10 +83,10 @@ BURN_UID = 111
 # to run on full dataset, set VOICEBENCH_MAX_SAMPLES to 100000, and make the dict empty.
 VOICEBENCH_MAX_SAMPLES=100000
 SAMPLES_PER_DATASET = {
-   'commoneval': 1, # 200
-   'wildvoice': 1, # 1000
-   'advbench': 1, # 520
-   'ifeval': 1 # 345
+   'commoneval': 50, # 200
+   'wildvoice': 100, # 1000
+   'advbench': 100, # 520
+   'ifeval': 50 # 345
 }
 
 # All available VoiceBench datasets with their appropriate splits. which datasets to run.
@@ -103,3 +103,5 @@ VOICEBENCH_DATASETS = {
     # 'bbh': ['test'],
     'advbench': ['test']
 }
+
+CONTAINER_TIMEOUT = 300  # seconds -> to prevent hanging containers
